+++ src/main/kotlin/ecommerce/repository/JdbcProductStore.kt
@@ -1,6 +1,6 @@
 package ecommerce.repository
 
-import ecommerce.model.Product
+import ecommerce.dto.Product
 import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.jdbc.core.RowMapper
 import org.springframework.stereotype.Repository
