+++ src/main/kotlin/ecommerce/validator/UniqueProductName.kt
@@ -4,7 +4,6 @@ import jakarta.validation.Constraint
 import jakarta.validation.Payload
 import kotlin.reflect.KClass
 
-@MustBeDocumented
 @Constraint(validatedBy = [UniqueProductNameValidator::class])
 @Target(AnnotationTarget.FIELD)
 @Retention(AnnotationRetention.RUNTIME)
